requirements.txt
fastapi
pytest
pylint
uvicorn[standard]
numpy
nest_asyncio
boto3
python-multipart
black
Pillow
ipython
nbconvert
requests